+++ packages/itwin/property-grid/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @itwin/property-grid-react
 
-This log was last generated on Mon, 19 May 2025 16:58:21 GMT and should not be manually modified.
+This log was last generated on Mon, 02 Jun 2025 13:19:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.19.1
+
+Mon, 02 Jun 2025 13:19:21 GMT
+
+### Patches
+
+- Fix property grid not loading data when using `SelectionStorage` directly. ([#1338](https://github.com/iTwin/viewer-components-react/pull/1338))
 
 ## 1.19.0